refactor: add routes
