+++ client/draft.md
@@ -1,0 +1,117 @@
+# 前端规划草稿
+
+
+## Country/area 200+个地区（由于路由，此处不再大写开头）
+
+- china                 中国
+- america               美国
+- korea                 韩国
+- uk                    英国
+- serbia                塞尔维亚
+- japan                 日本
+- singapore             新加坡
+- india                 印度
+- italy                 意大利
+- russia                俄罗斯
+- france                法国
+- spain                 西班牙
+- germany               德国
+- pakistan              巴基斯坦
+- iraq                  伊拉克
+- iran                  伊朗
+- laos                  老挝
+- cambodia              柬埔寨
+- philippines           菲律宾
+
+## Hot City 热门的城市、大型都市
+
+- wuhan                 武汉
+- newYork               纽约
+- tokyo                 东京
+- hongkong              香港
+- macao                 澳门
+- taiwan                台湾
+
+## loss                 损失
+
+- index                 概览
+- industry-assessment   行业评估。比如软件业可能影响较小的评估，可远程
+- economy               经济损失
+- car                   汽车行业损失
+- real-estate           房地产行业损失
+- company               企业损失
+- aquaculture           养殖业
+
+## ecology              生态
+- index                 概览
+- sea                   海洋生态
+- land                  陆地生态
+- antarctic             南极
+- arctic                北极
+- rivers                河流
+
+## Life                 声明、汇总一些名人科学家去世
+- scientist             科学家
+- star                  明星
+- politicians           政要
+- medical               医疗人员、医生
+- police                警察
+- businessmen           企业家
+
+## job                  就业：出现失业、就业难
+
+## travel               旅游行业
+
+- airline               航空
+- boat                  船舶
+
+## game                 游戏行业
+
+## entertainment        娱乐行业
+
+## research             研究
+- vaccine               疫苗
+- origin                COVID-19 where origin
+- paper                 论文，研究者发布的论文
+
+## rescue               医疗救援行动
+- to-italy              援助意大利
+- to-serbia             援助塞尔维亚
+
+## love                 捐助、捐款行动
+- index                 捐助概览
+- alibaba               阿里巴巴公益基金捐助
+...
+
+## global               全球
+- index                 全球概览
+  - total               感染确诊数目统计
+  - dead                死亡数目统计
+  - confirm             确诊数目增长和数目
+  - 统计多少个国家被感染
+  - 统计每个国家确诊数目的比例
+  - 统计每个国家的死亡数目
+  - 统计每个国家死亡率
+  - 统计每个国家治愈数目
+  - 统计每个国家现有治愈数目
+  - 统计每个国家的第一例从何时开始，做一些统计图
+  - 旅游高风险感染区
+  
+## who 世界卫生组织
+## solution
+- index                 解决方案概览
+  -- [covid-19-china-solution]() 发布中国解决方案的文档
+  
+## hero                 抗疫表现的备受鼓舞和信任的个人或团体
+- zhongnanshang         钟南山
+- lilanjuan             李兰娟
+- liwenliang            李文亮
+- zhangwenhong          张文宏
+- zhangboli             张伯礼
+
+
+## argue                争议
+
+## evaluate             评价
+
+## about                关于
