+++ client/package.json
@@ -6,6 +6,7 @@
   },
   "dependencies": {
     "@posva/vue-router-next": "^4.0.0-alpha.0",
+    "axios": "^0.19.2",
     "date-fns": "^2.9.0",
     "socket.io-client": "^2.3.0",
     "vue": "3.0.0-alpha.10",
@@ -14,6 +15,7 @@
   "devDependencies": {
     "@babel/core": "^7.5.5",
     "@babel/preset-env": "^7.5.5",
+    "@types/axios": "^0.14.0",
     "@vue/cli-plugin-babel": "^3.10.0",
     "@vue/compiler-sfc": "3.0.0-alpha.4",
     "babel-loader": "^7.1.5",
