+++ client/public/index.html
@@ -9,5 +9,5 @@
 </head>
 <div id="app"></div>
 <!--<script src="main.js"></script>-->
-<script src="bundle.js"></script>
+<script src="/bundle.js"></script>
 </html>
