--- client/src/components/modules/README.md
@@ -1,31 +1,0 @@
-# todo 业务组件
-- Dashboard/仪表盘
-
-- Map/地图图表
-
-- News/新闻
-
-- Timeline/时间轴
-仅有Console产生的人工确认数据发布到时间轴，以展示疫情整个生命周期
-
-- Post/发布
-
-- Console 控制台/授权中心
-产生入库的可信数据，人工操作
-
-- Chart/图表分析
-
-
-- API/API文档
-供给其他分析机构的API文档
-
-- AI/大数据分析/大数据研判
-
-- Data/数据衍生
- - 病毒类
- - 医学类
- 
- - Message/会话弹窗
- 
- - Barrage/弹幕消息
-
