--- client/src/config.js
@@ -1,2 +1,0 @@
-// export const ioServer = 'http://127.0.0.1:9999';
-export const ioServer = 'https://2019-ncov.datav.ai';
