+++ client/src/routes/aboutRoutes.ts
@@ -1,0 +1,15 @@
+/**
+ * @desc 关于
+ * @date 2020年4月3日11:51:05
+ * @author veaba
+ * */
+export const aboutRoutes = [
+    {
+        path: '/about',
+        children: [
+            {
+                path: '/index'
+            }
+        ]
+    }
+];
