+++ client/src/routes/argueRoutes.ts
@@ -1,0 +1,15 @@
+/**
+ * @desc 存在争议的问题
+ * @date 2020年4月3日11:50:15
+ * @author veaba
+ * */
+export const argueRoutes = [
+    {
+        path: '/argue',
+        children: [
+            {
+                path: '/index'
+            }
+        ]
+    }
+];
