+++ client/src/routes/ecologyRoutes.ts
@@ -1,0 +1,16 @@
+/**
+ * @desc 新冠对生态的影响
+ * @date 2020年4月3日11:42:28
+ * @author veaba
+ * */
+
+export const ecologyRoutes = [
+    {
+        path: '/ecology',
+        children: [
+            {
+                path:'/index'
+            }
+        ]
+    }
+];
