+++ client/src/routes/globalRoutes.ts
@@ -1,0 +1,15 @@
+/**
+ * @desc 全球概览
+ * @date 2020年4月3日11:48:26
+ * @author veaba
+ * */
+export const globalRoutes = [
+    {
+        path: '/global',
+        children: [
+            {
+                path:'/index'
+            }
+        ]
+    }
+];
