+++ client/src/routes/heroRoutes.ts
@@ -1,0 +1,15 @@
+/**
+ * @desc 抗疫表现的备受鼓舞和信任的个人或团体
+ * @date 2020年4月3日11:49:32
+ * @author veaba
+ * */
+export const heroRoutes = [
+    {
+        path: '/job',
+        children: [
+            {
+                path: '/index'
+            }
+        ]
+    }
+];
