+++ client/src/routes/jobRoutes.ts
@@ -1,0 +1,15 @@
+/**
+ * @desc 对就业的影响
+ * @date 2020年4月3日11:43:32
+ * @author veaba
+ * */
+export const jobRoutes = [
+    {
+        path: '/job',
+        children: [
+            {
+                path:'/index'
+            }
+        ]
+    }
+];
