+++ client/src/routes/lossRoutes.ts
@@ -1,0 +1,15 @@
+/**
+ * @desc 损失计算和评估
+ * @date 2020年4月3日11:38:55
+ * @author veaba
+ * */
+export const lossRoutes = [
+    {
+        path: '/loss',
+        children: [
+            {
+                path:'index'
+            }
+        ]
+    }
+];
