+++ client/src/routes/loveRoutes.ts
@@ -1,0 +1,18 @@
+/**
+ * @desc 捐助、捐款行动
+ * @date 2020年4月3日11:47:21
+ * @author veaba
+ * */
+export const loveRoutes = [
+    {
+        path: '/lovel',
+        children: [
+            {
+                path: 'index'
+            },
+            {
+                path: 'alibaba'
+            },
+        ]
+    }
+];
