+++ client/src/routes/whoRoutes.ts
@@ -1,0 +1,15 @@
+/**
+ * @desc WHO 世界卫生组织
+ * @date 2020年4月3日11:48:56
+ * @author veaba
+ * */
+export const whoRoutes = [
+    {
+        path: '/who',
+        children: [
+            {
+                path:'/index'
+            }
+        ]
+    }
+];
